refactor: 테스트 코드 문제 해결
